crypto-poker-webapp-main/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
aiogram
psycopg2-binary
